backend/requirements.txt
backports.csv==1.0.7
certifi==2019.9.11
chardet==3.0.4
defusedxml==0.6.0
diff-match-patch==20181111
dj-database-url==0.5.0
Django==2.2.6
django-allauth==0.40.0
django-appconf==1.0.3
django-cors-headers==3.1.1
django-crispy-forms==1.7.2
django-filter==2.2.0
django-formtools==2.1
django-heroku==0.3.1
django-imagekit==4.0.2
django-import-export==1.2.0
django-jet==1.0.8
django-rest-auth==0.9.5
django-reversion==3.0.4
djangorestframework==3.10.3
et-xmlfile==1.0.1
future==0.18.0
gunicorn==19.9.0
httplib2==0.9.2
idna==2.8
jdcal==1.4.1
Markdown==3.1.1
oauthlib==3.1.0
odfpy==1.4.0
openpyxl==3.0.0
pilkit==2.0
Pillow==6.2.0
psycopg2==2.8.3
python3-openid==3.1.0
pytz==2019.3
PyYAML==5.1.2
requests==2.22.0
requests-oauthlib==1.2.0
six==1.12.0
sorl-thumbnail==12.5.0
sqlparse==0.3.0
tablib==0.13.0
urllib3==1.25.6
WebOb==1.8.5
whitenoise==4.1.4
xadmin
xlrd
xlwt
